+++ src/libinstrum/common/Helpers.cpp
@@ -125,13 +125,13 @@ bool Helpers::writeFullFile(const std::string & fname, const std::string & data)
 
 	//write
 	ssize_t res = fwrite(data.c_str(), 1, data.size(), fp);
-	assert(res == data.size());
+	assert(res == static_cast<ssize_t>(data.size()));
 
 	//close
 	fclose(fp);
 
 	//ok
-	return res == data.size();
+	return res == static_cast<ssize_t>(data.size());
 }
 
 /**********************************************************/
